Change the start player's color in the settings dialog depending on who really is the start player.
Note: This directly sets the UI buttons in the KConfigDialog
which is not really nice and also requires the dialog to be modal.
However, KConfigDialog seems to not allow any other way of
changing its elements once it has been shown. Anybody knowing
a better solution feel free to improve it. 


svn path=/trunk/KDE/kdegames/kwin4/; revision=641209